requirements.txt
aiohue==4.3.0
aiohttp==3.8.1
attrs==21.4.0
click==8.0.4
jsonschema==4.4.0
paho-mqtt==1.6.1
pyyaml==6.0
tzlocal==4.1
